fix ugly bug in sample code
imagine if 2 handles finish in the same multi_exec(), before exec() $active is 2 and after the call $active is 0, then we read the first message, and the the message from the 2nd handle will be lost/ignored!

this has caused real bugs in the wild, see https://stackoverflow.com/questions/61920359/count-how-many-curl-multi-requests-have-been-made/61923372

Patch provided by divinity76.


git-svn-id: https://svn.php.net/repository/phpdoc/ja/trunk@349955 c90b9560-bf6c-de11-be94-00142212c4b1