filePro: extension was moved to PECL / no longer bundled with PHP as of PHP 5.2
As per: https://www.php.net/manual/en/intro.filepro.php

It is unclear to me what the current version nr is of the PECL extension, but showing the functions as if they are available in PHP > 5.2 seems wrong either way.

Includes adding a mention of when the extension was moved to PECL on the "Installing/Configuring" page.


git-svn-id: https://svn.php.net/repository/phpdoc/ja/trunk@349960 c90b9560-bf6c-de11-be94-00142212c4b1