Drop use of noise word "respectively" (#1165)
That's not what "respectively" means.

http://www.lexico.com/definition/respectively

"The PSpell functions now accept and return, respectively, PSpell\Config objects instead of pspell config resources."
means (a) there are precisely _two_ PSpell functions, where the first one (and you know which one that is) now accepts an object instead of a resource, and the second returns one.

https://github.com/php/doc-en/commit/33e1bb798e069750245a97c1cfa9bdc0f5cf2a02