disambiguate semantics of search_value
The definition of `search_value` in "Parameters" is a bit ambiguous. "these values" _could_ be interpreted to mean that if an array is passed as the `$search_value`, keys of elements of `$search_value` that match _any_ of those `$array` elements will be returned.
Note, the "Description" section does not have this ambiguity as it states "that value" instead of "those values".

https://github.com/php/doc-en/commit/8b00eb9807eb2aec3263226eb770a0885ebc804e