Add closure versions of all create_function examples
Basically the only reason people should be looking at the manual
for create_function right now is to see how *not* to use it. So,
for every example that was already included, I've added the
corresponding version using real anonymous functions.

Also, I've reworded the purpose and description, because these
functions aren't anonymous, as evidenced by the return value being
their name.

https://github.com/php/doc-en/commit/5065e44c26ccb29437ce26e9df0d26f446583d7a