Rewrite SoapClient constructor documentation
This constructor takes a large number of options, most of which
were effectively undocumented, leading to a very high volume of
User Notes. Some examples and caveats from these notes have been
incorporated.

The descriptions here are largely based on reverse-engineering
the source code (which isn't particularly well commented) and
experimenting locally, so may not be 100% accurate, or capture
differences between versions.

今更 SOAP のドキュメントを全面的に書き直されても、使う人いるのかしら...

https://github.com/php/doc-en/commit/6105da168f7d377a7b122b83e25ddac6b5dd2191