Fix XMLParser example for more recent PHP versions
The example uses the `$parser` as array index, but that implicit
conversion to integer already raises a notice as of PHP 7.0.0, and
outright fails as of PHP 8.0.0, since the `$parser` is an object now.

Since there doesn't appear a good reason for `$depth` to be an array
which supports different parsers, we change it to an integer, what also
simplifies the example.

This basically integrates user note 127199.

https://github.com/php/doc-en/commit/f5992156e664c34794804c282a95f6428a2687a7