Change simplesect to sect1 in exception chapter
This will fix the fact that the examples start their number at 3.
The reason for that is the numbering starts within the previous ``<sect1>`` but this is not rendered on the page, but in a follow-up page.
However, we cannot put this sections after the ``<simplesect>`` as that doesn't respect the DTD: https://tdg.docbook.org/tdg/5.0/chapter.html

Converting them to <sect1>, however, makes them render into different pages.
Therefore, we use our custom PhD DTD: ``xmlns:phd="http://www.php.net/ns/phd"`` and use the ``<sect1 phd:chunk="false">`` tag which will keep the current behaviour of not splitting the pages.

日本語版を追随させるにあたっては、以下の修正も追加している

- 「スローする」に統一
- その他細かい語尾の修正

https://github.com/php/doc-en/commit/491a7cc9cddf1e20cda76b16226d0bc020402476