Better crosslinks between str_replace and preg_replace
The str_replace page had a helpful note that if you'd somehow got
there when you were looking for preg_replace, you didn't need to
llok further. But if you were actually on the page for preg_replace,
there was no link to str_replace at all, not even in the See Also
section.

https://github.com/php/doc-en/commit/464fbf0d12f89b1e723908780fb68acf667f5901