Remove superfluous changelog entries
Actually, most (or even all) datetime functions returned false prior to
PHP 8.0.0, where they now throw an error.  Besides ZPP failures which
we certainly do not want to document explicitly, that also happened in
case the objects have not been properly initialized, which is so close
to ZPP failures, that we don't want to document that either.

https://github.com/php/doc-en/commit/e057f1f552548a68030830e96ae1bdf313b1794e