ref the exact section of ISO8601 broken
historically there has been much confusion about whether or not DateTime::ISO8601 actually breaks the ISO8601 specs or not.

as late as an early draft of "Deprecations for PHP 8.1" RFC incorrectly contained

>> There's a number of bug reports related to this. From what I understand, the core problem here is not that the ISO8601 format is wrong, it's just one of multiple legal ISO-8601 formats. As DateTime formats always refer to a specific format, not a set of multiple possible ones, there doesn't seem to be anything actionable here.

so it's probably a good idea to add a reference to the exact part of ISO8601 being broken.

https://archive.org/details/iso-tc154-wg5_n0038_iso_wd_8601-1_2016-02-16/page/20/mode/2up

https://github.com/php/doc-en/commit/cb5790207359b1e69d98ec95d37f8fb1061ffec8