Make password_needs_rehash() more prominent
Successful validation should usually be followed by re-hashing the
password, if necessary.  The `password_needs_rehash()` docs already
have a good example, but the `password_verify()` page does not mention
that at all.  Instead of duplicating the example, we refer to it.

We also remove the link to the password compatibility library, since it
is no longer relevant for the manual.

https://github.com/php/doc-en/commit/abc0c909d5a642fbc9008c5ed1a4c3ead5ecb9a2