random: Remove examples for mt_srand() and srand()
These examples teach the readers bad practices, because they indicate that
seeding from the current time is reasonable (it is not). Users should ideally
use the `random_int()` function. If that is not possible, the
`Random\Randomizer` is the next best choice. If that's not an option either,
then relying on the random seed of the global Mt19937 is better than any manual
seeding that relies on the current time.

https://github.com/php/doc-en/commit/823cfabfbd98d69e7bae993c5edc6e92a20bab77