stream_get_contents: Note buffer of max length size will always be allocated
Clarify in a note that even if a stream's contents are much smaller, a buffer of the maximum length will be allocated if a maximum length is provided. This may be surprising if you otherwise expect the function to dynamically grow a buffer depending on the actual stream contents, as it does with the default null value.

https://github.com/php/doc-en/commit/d5ed467fdff1fc6c6d2c4af41a7e7d6d95c25dd2